refactor: project services to functions
