+++ src/app/resume-page/projects-section/projects-section.component.spec.ts
@@ -1,13 +1,14 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing'
+import { ComponentFixture } from '@angular/core/testing'
 
 import { ProjectsSectionComponent } from './projects-section.component'
-import { MockComponents } from 'ng-mocks'
+import { MockComponents, MockProvider } from 'ng-mocks'
 import { SectionTitleComponent } from '../section-title/section-title.component'
 import { byComponent } from '@/test/helpers/component-query-predicates'
-import { ProjectItemsService } from './project-items.service'
+import { GET_PROJECT_ITEMS, GetProjectItems } from './get-project-items'
 import { ProjectItemComponent } from './project-item/project-item.component'
 import { NgFor } from '@angular/common'
 import { componentTestSetup } from '@/test/helpers/component-test-setup'
+import { makeProjectItem } from './project-item/__tests__/make-project-item'
 
 describe('ProjectsSectionComponent', () => {
   let component: ProjectsSectionComponent
@@ -19,23 +20,30 @@ describe('ProjectsSectionComponent', () => {
   })
 
   it('should display all projects', () => {
-    ;[fixture, component] = makeSut()
+    const projectItems = [makeProjectItem(), makeProjectItem()]
+    const getProjectItems = jasmine
+      .createSpy<GetProjectItems>()
+      .and.returnValues(projectItems)
+    ;[fixture, component] = makeSut({ getProjectItems })
     fixture.detectChanges()
 
-    const projectItemsService = TestBed.inject(ProjectItemsService)
     const projectItemElements = fixture.debugElement.queryAll(
       byComponent(ProjectItemComponent),
     )
-    expect(projectItemElements.length).toEqual(projectItemsService.get().length)
+    expect(projectItemElements.length).toEqual(projectItems.length)
   })
 })
 
-function makeSut() {
-  return componentTestSetup(ProjectsSectionComponent, {
+const makeSut = (opts: { getProjectItems?: GetProjectItems } = {}) =>
+  componentTestSetup(ProjectsSectionComponent, {
     imports: [
       ProjectsSectionComponent,
       NgFor,
       MockComponents(SectionTitleComponent, ProjectItemComponent),
     ],
+    providers: [
+      opts.getProjectItems
+        ? MockProvider(GET_PROJECT_ITEMS, opts.getProjectItems)
+        : [],
+    ],
   })
-}
